+++ app/Http/Controllers/Admin/SalaryTypeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\SalaryType;
 use App\Services\Notify;
 use App\Traits\Searchable;
@@ -93,6 +94,11 @@ class SalaryTypeController extends Controller
      */
     public function destroy(string $id): Response
     {
+        $jobExist = Job::where('salary_type_id', $id)->exists();
+        if ($jobExist) {
+            return response(['message' => 'This item is already been used can\'t delete! 🚫'], 500);
+        }
+
         try {
             SalaryType::findOrFail($id)->delete();
             Notify::deletedNotification();
